Fix FlexLexer

+++ src/parser/sql-stmt/Literal.hpp
@@ -21,7 +21,7 @@ class Literal : public Expression
 public:
   Literal() : Expression(LITERAL), valueType_(LiteralType::NONE) {}
   explicit Literal(bool v) : Expression(LITERAL), valueType_(LiteralType::BOOL), ival_(v) {}
-  explicit Literal(int64_t v) : Expression(LITERAL), valueType_(LiteralType::INTEGER), fval_(v) {}
+  explicit Literal(long long v) : Expression(LITERAL), valueType_(LiteralType::INTEGER), fval_(v) {}
   explicit Literal(long double v) : Expression(LITERAL), valueType_(LiteralType::DOUBLE), bval_(v) {}
   explicit Literal(std::string v) : Expression(LITERAL), valueType_(LiteralType::TEXT), strval_(v) {}
   
@@ -47,7 +47,7 @@ private:
 
   bool bval_;
   double fval_;
-  int64_t ival_;
+  long long ival_;
   std::string strval_;
 };
